Make HTTP caching more testable by splitting implementation out of HttpSource Add NUGET_HTTP_CACHE_PATH environment variable for setting the HTTP cache path
