+++ src/NuGet.Core/NuGet.Protocol.Core.v3/HttpSource/HttpSource.cs
@@ -4,11 +4,8 @@
 using System;
 using System.Globalization;
 using System.IO;
-using System.Linq;
 using System.Net;
 using System.Net.Http;
-using System.Security.Cryptography;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
@@ -21,7 +18,6 @@ namespace NuGet.Protocol
 {
     public class HttpSource : IDisposable
     {
-        private const int BufferSize = 8192;
         private readonly Func<Task<HttpHandlerResource>> _messageHandlerFactory;
         private readonly Uri _baseUri;
         private HttpClient _httpClient;
@@ -70,7 +66,11 @@ namespace NuGet.Protocol
             ILogger log,
             CancellationToken token)
         {
-            var result = InitializeHttpCacheResult(request.CacheKey, request.CacheContext);
+            var result = HttpCacheUtility.InitializeHttpCacheResult(
+                HttpCacheDirectory,
+                _baseUri,
+                request.CacheKey,
+                request.CacheContext);
 
             return await ConcurrencyUtilities.ExecuteWithFileLockedAsync(
                 result.CacheFile,
@@ -140,7 +140,7 @@ namespace NuGet.Protocol
 
                         throttledResponse.Response.EnsureSuccessStatusCode();
 
-                        await CreateCacheFileAsync(
+                        await HttpCacheUtility.CreateCacheFileAsync(
                             result,
                             request.Uri,
                             throttledResponse.Response,
@@ -300,103 +300,13 @@ namespace NuGet.Protocol
             return httpClient;
         }
 
-        private HttpCacheResult InitializeHttpCacheResult(string cacheKey, HttpSourceCacheContext context)
-        {
-            // When the MaxAge is TimeSpan.Zero, this means the caller is passing in a folder different than
-            // the global HTTP cache used by default. Additionally, creating and cleaning up the directory is
-            // all the responsibility of the caller.
-            var maxAge = context.MaxAge;
-            string newFile;
-            string cacheFile;
-            if (!maxAge.Equals(TimeSpan.Zero))
-            {
-                var baseFolderName = RemoveInvalidFileNameChars(ComputeHash(_baseUri.OriginalString));
-                var baseFileName = RemoveInvalidFileNameChars(cacheKey) + ".dat";
-
-                var cacheFolder = Path.Combine(HttpCacheDirectory, baseFolderName);
-
-                cacheFile = Path.Combine(cacheFolder, baseFileName);
-
-                newFile = cacheFile + "-new";
-            }
-            else
-            {
-                cacheFile = Path.Combine(context.RootTempFolder, Path.GetRandomFileName());
-
-                newFile = Path.Combine(context.RootTempFolder, Path.GetRandomFileName());
-            }
-
-            return new HttpCacheResult(maxAge, newFile, cacheFile);
-        }
-
-        private async Task CreateCacheFileAsync(
-            HttpCacheResult result,
-            string uri,
-            HttpResponseMessage response,
-            HttpSourceCacheContext context,
-            Action<Stream> ensureValidContents,
-            CancellationToken cancellationToken)
-        {
-            // The update of a cached file is divided into two steps:
-            // 1) Delete the old file.
-            // 2) Create a new file with the same name.
-            using (var fileStream = new FileStream(
-                result.NewCacheFile,
-                FileMode.Create,
-                FileAccess.ReadWrite,
-                FileShare.None,
-                BufferSize,
-                useAsync: true))
-            {
-                using (var networkStream = await response.Content.ReadAsStreamAsync())
-                {
-                    await networkStream.CopyToAsync(fileStream, 8192, cancellationToken);
-                }
-
-                // Validate the content before putting it into the cache.
-                fileStream.Seek(0, SeekOrigin.Begin);
-                ensureValidContents?.Invoke(fileStream);
-            }
-
-            if (File.Exists(result.CacheFile))
-            {
-                // Process B can perform deletion on an opened file if the file is opened by process A
-                // with FileShare.Delete flag. However, the file won't be actually deleted until A close it.
-                // This special feature can cause race condition, so we never delete an opened file.
-                if (!IsFileAlreadyOpen(result.CacheFile))
-                {
-                    File.Delete(result.CacheFile);
-                }
-            }
-
-            // If the destination file doesn't exist, we can safely perform moving operation.
-            // Otherwise, moving operation will fail.
-            if (!File.Exists(result.CacheFile))
-            {
-                File.Move(
-                    result.NewCacheFile,
-                    result.CacheFile);
-            }
-
-            // Even the file deletion operation above succeeds but the file is not actually deleted,
-            // we can still safely read it because it means that some other process just updated it
-            // and we don't need to update it with the same content again.
-            result.Stream = new FileStream(
-                result.CacheFile,
-                FileMode.Open,
-                FileAccess.Read,
-                FileShare.Read | FileShare.Delete,
-                BufferSize,
-                useAsync: true);
-        }
-
         public string HttpCacheDirectory
         {
             get
             {
                 if (_httpCacheDirectory == null)
                 {
-                    _httpCacheDirectory = NuGetEnvironment.GetFolderPath(NuGetFolderPath.HttpCacheDirectory);
+                    _httpCacheDirectory = SettingsUtility.GetHttpCacheFolder();
                 }
 
                 return _httpCacheDirectory;
@@ -407,80 +317,7 @@ namespace NuGet.Protocol
 
         protected virtual Stream TryReadCacheFile(string uri, TimeSpan maxAge, string cacheFile)
         {
-            if (!maxAge.Equals(TimeSpan.Zero))
-            {
-                string cacheFolder = Path.GetDirectoryName(cacheFile);
-                if (!Directory.Exists(cacheFolder))
-                {
-                    Directory.CreateDirectory(cacheFolder);
-                }
-            }
-
-            if (File.Exists(cacheFile))
-            {
-                var fileInfo = new FileInfo(cacheFile);
-                var age = DateTime.UtcNow.Subtract(fileInfo.LastWriteTimeUtc);
-                if (age < maxAge)
-                {
-                    var stream = new FileStream(
-                        cacheFile,
-                        FileMode.Open,
-                        FileAccess.Read,
-                        FileShare.Read | FileShare.Delete,
-                        BufferSize,
-                        useAsync: true);
-
-                    return stream;
-                }
-            }
-
-            return null;
-        }
-
-        private static string ComputeHash(string value)
-        {
-            var trailing = value.Length > 32 ? value.Substring(value.Length - 32) : value;
-            byte[] hash;
-            using (var sha = SHA1.Create())
-            {
-                hash = sha.ComputeHash(Encoding.UTF8.GetBytes(value));
-            }
-
-            const string hex = "0123456789abcdef";
-            return hash.Aggregate("$" + trailing, (result, ch) => "" + hex[ch / 0x10] + hex[ch % 0x10] + result);
-        }
-
-        private static string RemoveInvalidFileNameChars(string value)
-        {
-            var invalid = Path.GetInvalidFileNameChars();
-            return new string(
-                value.Select(ch => invalid.Contains(ch) ? '_' : ch).ToArray()
-                )
-                .Replace("__", "_")
-                .Replace("__", "_");
-        }
-
-        private static bool IsFileAlreadyOpen(string filePath)
-        {
-            FileStream stream = null;
-
-            try
-            {
-                stream = new FileStream(filePath, FileMode.Open, FileAccess.ReadWrite, FileShare.None);
-            }
-            catch
-            {
-                return true;
-            }
-            finally
-            {
-                if (stream != null)
-                {
-                    stream.Dispose();
-                }
-            }
-
-            return false;
+            return HttpCacheUtility.TryReadCacheFile(uri, maxAge, cacheFile);
         }
 
         public static HttpSource Create(SourceRepository source)
@@ -548,21 +385,6 @@ namespace NuGet.Protocol
                     Interlocked.Exchange(ref _throttle, null)?.Release();
                 }
             }
-        }
-
-        private class HttpCacheResult
-        {
-            public HttpCacheResult(TimeSpan maxAge, string newCacheFile, string cacheFile)
-            {
-                MaxAge = maxAge;
-                NewCacheFile = newCacheFile;
-                CacheFile = cacheFile;
-            }
-
-            public TimeSpan MaxAge { get; }
-            public string NewCacheFile { get; }
-            public string CacheFile { get; }
-            public Stream Stream { get; set; }
         }
     }
 }
